+++ fava/core/__init__.py
@@ -18,17 +18,17 @@ from beancount.utils.encryption import is_encrypted_file
 from beancount.utils.misc_utils import filter_type
 
 from fava.util import date
-from fava.api.budgets import BudgetModule
-from fava.api.charts import ChartModule
-from fava.api.fava_options import parse_options
-from fava.api.file import FileModule
-from fava.api.filters import (AccountFilter, FromFilter, PayeeFilter,
-                              TagFilter, TimeFilter)
-from fava.api.helpers import FavaAPIException, FavaModule, entry_at_lineno
-from fava.api.holdings import get_final_holdings, aggregate_holdings_by
-from fava.api.misc import FavaMisc
-from fava.api.query_shell import QueryShell
-from fava.api.watcher import Watcher
+from fava.core.budgets import BudgetModule
+from fava.core.charts import ChartModule
+from fava.core.fava_options import parse_options
+from fava.core.file import FileModule
+from fava.core.filters import (AccountFilter, FromFilter, PayeeFilter,
+                               TagFilter, TimeFilter)
+from fava.core.helpers import FavaAPIException, FavaModule, entry_at_lineno
+from fava.core.holdings import get_final_holdings, aggregate_holdings_by
+from fava.core.misc import FavaMisc
+from fava.core.query_shell import QueryShell
+from fava.core.watcher import Watcher
 
 
 def _list_accounts(root_account, active_only=False):
