+++ fava/core/misc.py
@@ -5,7 +5,7 @@ import datetime
 from beancount.core.data import Custom, Event
 from beancount.utils.misc_utils import filter_type
 
-from fava.api.helpers import FavaModule
+from fava.core.helpers import FavaModule
 
 
 # pylint: disable=too-few-public-methods
